WIP: Remove deprecated balena-errors
I basically replicated some code from `balena-errors`, including tests

I believe this approach will result in the same issue that we are
trying to avoid deprecating `balena-errors` when trying to
check errors using `error instanceof BalenaError`. So I'm not sure if
that's the way that we should proceed.

We could keep `balena-errors` only for a reusable definition of
`BalenaError`.
That would hardly change and remove some of the replication.

A `BalenaError` instance and `exitCode` are expected on
[balena-cli::errors](https://github.com/balena-io/balena-cli/blob/2cc8d15c0535ef90826edf9746bab3b03cb95d0f/lib/errors.ts#L178)